rename trestle.core.commands.author.profile to prof - lint shadow issue
