+++ .pylintrc_tests
@@ -3,7 +3,7 @@ ignore=oscal
 
 
 [MESSAGES CONTROL]
-disable=W1203, W1201, W0212, A005
+disable=W1203, W1201, W0212
 
 [FORMAT]
 max-line-length=120
