+++ tests/trestle/core/commands/author/prof_test.py
@@ -38,7 +38,7 @@ from trestle.common.err import TrestleError
 from trestle.common.list_utils import comma_colon_sep_to_dict, comma_sep_to_list
 from trestle.common.model_utils import ModelUtils
 from trestle.core.catalog.catalog_interface import CatalogInterface
-from trestle.core.commands.author.profile import ProfileAssemble, ProfileGenerate, ProfileInherit
+from trestle.core.commands.author.prof import ProfileAssemble, ProfileGenerate, ProfileInherit
 from trestle.core.control_interface import ControlInterface
 from trestle.core.markdown.docs_markdown_node import DocsMarkdownNode
 from trestle.core.markdown.markdown_api import MarkdownAPI
