+++ trestle/core/commands/author/command.py
@@ -26,7 +26,7 @@ from trestle.core.commands.author.docs import Docs
 from trestle.core.commands.author.folders import Folders
 from trestle.core.commands.author.headers import Headers
 from trestle.core.commands.author.jinja import JinjaCmd
-from trestle.core.commands.author.profile import ProfileAssemble, ProfileGenerate, ProfileInherit, ProfileResolve
+from trestle.core.commands.author.prof import ProfileAssemble, ProfileGenerate, ProfileInherit, ProfileResolve
 from trestle.core.commands.author.ssp import SSPAssemble, SSPFilter, SSPGenerate
 from trestle.core.commands.command_docs import CommandPlusDocs
